♻️  Introduce on function for all providers (canConnectWith)
