--- packages/applications/request-context/src/canConnectWithMagicLink.ts
@@ -1,11 +1,0 @@
-import { PlainType } from '@potentiel-domain/core';
-import { Role } from '@potentiel-domain/utilisateur';
-
-export const canConnectWithMagicLink = (role: PlainType<Role.ValueType>) => {
-  const roleValueType = Role.bind(role);
-  const admin = roleValueType.estÉgaleÀ(Role.admin);
-  const dgecval = roleValueType.estÉgaleÀ(Role.dgecValidateur);
-  const dreal = roleValueType.estÉgaleÀ(Role.dreal);
-
-  return !(admin || dgecval || dreal);
-};
